requirements.txt
chromadb==0.6.1
fastapi[standard]==0.115.6
langchain-core==0.3.29
langchain-ollama==0.2.2
langchain-openai==0.2.14
nltk==3.9.1
ollama==0.4.5
openai==1.59.3
python-dotenv==1.0.1
scikit-learn==1.6.0
sqlalchemy==2.0.36
uuid-shortener==0.1.3
